### 2026-04-02 #10364 @editorialbot
[PRE REVIEW]: MooBench: A micro-benchmark for performance overhead measurement of observability tools
**Submitting author:** @DaGeRe (<a href="http://orcid.org/0000-0002-1772-1416">David Georg Reichelt</a>)
**Repository:** https://github.com/kieker-monitoring/moobench
**Branch with paper.md** (empty if default branch): moobench-paper
**Version:** v1.0.1
