### 2026-04-02 ac5207e
Add missing doi instrumentation frameworks
