+++ moobench-paper/paper.bib
@@ -294,7 +294,8 @@ series = {ICPE '25}
   author={Reichelt, David Georg and Bulej, Lubom{\'\i}r and Jung, Reiner and Van Hoorn, Andr{\'e}},
   booktitle={Companion of the 15th ACM/SPEC International Conference on Performance Engineering},
   pages={249--256},
-  year={2024}
+  year={2024},
+  doi={10.1145/3629527.3652269}
 }
 
 @article{reichelt2021overhead,
