### 2026-04-02 d9defee
Display doi for Kieker paper
