+++ moobench-paper/paper.bib
@@ -366,7 +366,7 @@ author = "Wilhelm Hasselbring and André {van Hoorn}"
   year		= {2012},
   pages	= {247--248},
   url		= {http://eprints.uni-kiel.de/14418/},
-  Xdoi		= {10.1145/2188286.2188326},
+  doi		= {10.1145/2188286.2188326},
   abstract	= {Kieker is an extensible framework for monitoring and
 		  analyzing the runtime behavior of concurrent or distributed
 		  software systems. It provides measurement probes for
