### 2026-04-02 #10364 @danielskatz
@DaGeRe - Please fix the invalid DOIs mentioned above.  Note that https://doi.org/10.1145/3777884.3797004 and https://doi.org/10.1109/MCSE.2026.3662117 don't resolve.
