### 2026-04-02 #10364 @danielskatz
@DaGeRe - Also in your paper.md, when you refer to a reference directly, you should remove the []s around it, unlike when you use it parenthetically to support other text.
