### 2026-04-02 f08bc04
Fix citation style
