+++ moobench-paper/paper.md
@@ -36,7 +36,7 @@ aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
 
 Understanding the runtime behavior of software is inherently difficult due to the unpredictability of the software's behavior itself and the non-determinism of underlying layers, such as Just-In-Time (JIT) compilation in virtual machines, operating system scheduling, and CPU frequency scaling. Observability tools aim to answer questions regarding runtime behavior of software, such as "How much time did this request take?" or "How often did method A call method B?" [@majors2022observability]. These questions are answered using telemetry data, i.e., measurement data that is obtained from the code execution. To collect telemetry data, additional code needs to be executed, which introduces overhead. This overhead affects both system performance and the accuracy of the measurements themselves. The MooBench microbenchmark measures this overhead and contains factorial experiments that facilitate breaking down this overhead into its root causes.
 
-The MooBench benchmark was originally developed to examine the performance overhead of the Kieker observability framwork [@kieker2012] [@yang2025kieker] [@Kieker2020] in Java [@waller2013benchmark] and was extended as a general overhead measurement microbenchmark for various observability tools, currently within the Java and Python ecosystem. In this paper, we describe why it is needed, how it is structured, and how it is used in research.
+The MooBench benchmark was originally developed to examine the performance overhead of the Kieker observability framwork [@kieker2012; @yang2025kieker; @Kieker2020] in Java [@waller2013benchmark] and was extended as a general overhead measurement microbenchmark for various observability tools, currently within the Java and Python ecosystem. In this paper, we describe why it is needed, how it is structured, and how it is used in research.
 
 # Statement of need
 
@@ -44,9 +44,9 @@ Observability tools are crucial for managing software system health and optimizi
 
 # State of the field
 
-In the context of microservices, various studies have examined the overhead of tracing. For example, Nou et al. [@nou2025investigating] investigated the overhead of OpenTelemetry and Elastic APM using open-source applications, and examined the root causes of the overhead using profiling. Ahmed et al. [@ahmed2016studying] compared the overhead of commercial observability tools and the open-source tool Pinpoint for regression detection. Eder et al. [@eder2023comparison] compared the overhead using serverless applications.
+In the context of microservices, various studies have examined the overhead of tracing. For example, @nou2025investigating investigated the overhead of OpenTelemetry and Elastic APM using open-source applications, and examined the root causes of the overhead using profiling. @ahmed2016studying compared the overhead of commercial observability tools and the open-source tool Pinpoint for regression detection. @eder2023comparison compared the overhead using serverless applications.
 
-Besides the microservice context, tracing overhead has been examined for operating systems and High-Performance Computing (HPC). For operating systems, the overhead of tools for tracing the kernel itself [@desnoyers2006lttng] [@gebai2018tracers] and eBPF, that allows to trace both within the kernel and the user space [@volpertEBPFOverhead] [@domaschka2023using], has been examined in case studies. In the context of HPC, research has addressed the impact of overhead [@hunold2022overhead] and the combination of instrumentation and sampling to reduce it [@ilsche2015combining].
+Besides the microservice context, tracing overhead has been examined for operating systems and High-Performance Computing (HPC). For operating systems, the overhead of tools for tracing the kernel itself [@desnoyers2006lttng; @gebai2018tracers] and eBPF, that allows to trace both within the kernel and the user space [@volpertEBPFOverhead; @domaschka2023using], has been examined in case studies. In the context of HPC, research has addressed the impact of overhead [@hunold2022overhead] and the combination of instrumentation and sampling to reduce it [@ilsche2015combining].
 
[…]
