### 2026-04-02 #10364 @danielskatz
> Thanks for the hint, I've removed the [] for direct references.
>
> I can't resolve the DOI issue directly, since the papers have been accepted and DOIs have been registered, but not published yet (for the ICPE, the paper is also listed as accepted paper: https://icpe2026.spec.org/accepted-papers/ ). Should I remove the DOIs, or just leave it as it is?

I guess you should leave them as is for now, but if they don't work by the time the paper is ready to be published, we might have to reconsider this.
