### 2026-04-07 #10364 @jatkinson1000
Hi @danielskatz it was a long (4-day) weekend for Easter in the UK, so just following up on things today.
Yes, I am happy to take this on - ~perhaps slightly outside my usual track~ (on closer reading it's actually fine) it's an RSECon submission, am I correct?

@DaGeRe thank you for the submission and resolving the above issues. I'll take a read of the paper and begin searching for reviewers, but this may be Friday when I have time set aside for JOSS as I am presenting at a conference Thursday.
