### 2026-04-08 11eceec
Fix README: List all fully supported frameworks, fix formatting of enumerations
