### 2026-04-09 02c5646
Add Kieker variant for CLOCK_REALTIME timer
