+++ frameworks/Kieker-java/labels.sh
@@ -4,3 +4,4 @@ TITLE[2]="No collection"
 TITLE[3]="Text file"
 TITLE[4]="Binary file"
 TITLE[5]="Binary TCP"
+TITLE[6]="Binary file (Milli timer)"
