### 2026-04-09 677a2ca
Downgrade skywalking-java to Banyandb 0.9.0
