### 2026-04-09 02b3578
Check whether the given folder is absolute
