### 2026-04-10 #10364 @jatkinson1000
Hi there @akshaymittal143, @sampottinger, @Ahanaf-Aziz, @shrishar, and @abhishektiwari!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a framework for measuring the overheads introduced by adding observability/telemery features to Java and Python codes such that these might be accounted for in the overall benchmarks.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
