### 2026-04-10 #10364 @shrishar
Yes happy to review

On Fri, Apr 10, 2026 at 4:28 AM The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* left a comment (openjournals/joss-reviews#10364)
> <https://github.com/openjournals/joss-reviews/issues/10364#issuecomment-4223435026>
>
> @Ahanaf-Aziz <https://github.com/Ahanaf-Aziz> added to the reviewers list!
>
