### 2026-04-11 de4cbbb
Remove no longer maintained SPASSmeter framework
