+++ frameworks/SPASSmeter/benchmark-tcp.sh
@@ -1,141 +1,0 @@
-#!/bin/bash
-# This file is configured for linux instead of solaris!!!
-
-JAVABIN=""
-
-R_SCRIPT_DIR=r/
-BASE_DIR=./
-RESULTS_DIR="${BASE_DIR}/tmp/results-spassmeter/"
-
-SLEEP_TIME=30           ## 30
-NUM_LOOPS=10           ## 10
-THREADS=1              ## 1
-RECURSION_DEPTH=10      ## 10
-TOTAL_CALLS=2000000     ## 2000000
-METHOD_TIME=0      ## 500000
-
-#MORE_PARAMS="--quickstart"
-MORE_PARAMS="--application mooBench.monitoredApplication.MonitoredClassSimple ${MORE_PARAMS}"
-
-TIME=`expr ${METHOD_TIME} \* ${TOTAL_CALLS} / 1000000000 \* 4 \* ${RECURSION_DEPTH} \* ${NUM_LOOPS} + ${SLEEP_TIME} \* 4 \* ${NUM_LOOPS}  \* ${RECURSION_DEPTH} + 50 \* ${TOTAL_CALLS} / 1000000000 \* 4 \* ${RECURSION_DEPTH} \* ${NUM_LOOPS} `
-echo "Experiment will take circa ${TIME} seconds."
-
-echo "Removing and recreating '${RESULTS_DIR}'"
-(rm -rf ${RESULTS_DIR}/) && mkdir ${RESULTS_DIR}/
-#mkdir ${RESULTS_DIR}/stat/
-
-# Clear spassmeter.log and initialize logging
-rm -f ${BASE_DIR}/spassmeter.log
-touch ${BASE_DIR}/spassmeter.log
-
-RAWFN="${RESULTS_DIR}/raw"
-
-JAVA_ARGS="-server"
-JAVA_ARGS="${JAVA_ARGS} -d64"
-JAVA_ARGS="${JAVA_ARGS} -Xms1G -Xmx4G"
-JAVA_ARGS="${JAVA_ARGS} -verbose:gc -XX:+PrintCompilation"
-#JAVA_ARGS="${JAVA_ARGS} -XX:+PrintInlining"
-#JAVA_ARGS="${JAVA_ARGS} -XX:+UnlockDiagnosticVMOptions -XX:+LogCompilation"
-#JAVA_ARGS="${JAVA_ARGS} -Djava.compiler=NONE"
-JAR="-jar MooBench.jar"
-
-JAVA_ARGS_NOINSTR="${JAVA_ARGS}"
-JAVA_ARGS_LTW="${JAVA_ARGS} -javaagent:${BASE_DIR}/lib/linux/spass-meter-ia.jar=xmlconfig=${BASE_DIR}/lib/config.xml,out=${RESULTS_DIR}/spassmeter.txt,tcp=localhost:6002"
-JAVA_ARGS_LTW_ASM="${JAVA_ARGS_LTW} -Dspass-meter.iFactory=de.uni_hildesheim.sse.monitoring.runtime.instrumentation.asmTree.Factory"
-
-SERVER="${JAVA_ARGS} -classpath ${BASE_DIR}/lib/linux/spass-meter-ant.jar de.uni_hildesheim.sse.monitoring.runtime.recordingServer.TCPRecordingServer baseDir=. port=6002"
-
-## Write configuration
-uname -a >${RESULTS_DIR}/configuration.txt
-${JAVA_BIN} ${JAVA_ARGS} -version 2>>${RESULTS_DIR}/configuration.txt
-echo "JAVA_ARGS: ${JAVA_ARGS}" >>${RESULTS_DIR}/configuration.txt
-echo "" >>${RESULTS_DIR}/configuration.txt
[…]
