+++ README.md
@@ -102,12 +102,12 @@ cd frameworks/OpenTelemetry-java/
 - benchmark = moobench code
 - continuous-integration = obsolete
 - docker = Dockerfile to be used in automated runs on our intrestructure
-- frameworks = benchmark setups for the different frameworks for the respective language
+- frameworks = benchmark setups for the different frameworks for the respective language (in the form $FRAMEWORK-$LANGUAGE)
   - Kieker-java
   - Kieker-python
   - OpenTelementry-java
-  - SPASSmeter (currently not maintained)
   - inspectIT-java
+  - ...
 - gradle = build system, gradle libraries
 - tools = tooling used to support benchmarks and process results
   - compile-results = adds new results to a result log and computes partial

[…]
