### 2026-04-16 kieker-monitoring/moobench#49 @abhishektiwari
[JOSS Review] Improve Documentation/README
- Add a section calling out prerequisite for installation E.g., Java, R, Gradle, etc. Provide supported versions, OS, etc.

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/10400)
