### 2026-04-16 kieker-monitoring/moobench#50 @abhishektiwari
[JOSS Review] Provide working example
* Provide Example usage. Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

From README it's unclear if there is a working example in repository.

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/10400)
