### 2026-04-16 babdb0e
Split prerequisites and have clearer benchmark execution description
