### 2026-04-16 kieker-monitoring/moobench#50 @DaGeRe
Thanks for the issue.

I've tried to clarify this in the README: https://github.com/kieker-monitoring/moobench?tab=readme-ov-file#benchmark-execution-1 Does this work for you?
