### 2026-04-16 kieker-monitoring/moobench#49 @DaGeRe
Thanks for the hint, I've added a section: https://github.com/kieker-monitoring/moobench?tab=readme-ov-file#benchmark-execution-1 Is this sufficient?
(Gradle doesn't need to be installed, since the Gradle Wrapper should be used)
