### 2026-04-17 #10364 @jatkinson1000
Hi @akshaymittal143 We now have three reviewers in place (JOSS suggests 3, minimum 2) so the review has started.
If we lose any reviewers I'll be sure to come back to you, and will bear you in mind for similar papers in future, but for now we are all OK thanks!
