### 2026-04-18 kieker-monitoring/moobench#51 @shrishar
[JOSS Review] Add Example Usage on how do one use software real word analysis
Add few examples on how to use the software in the README.
[JOSS REVIEW](https://github.com/openjournals/joss-reviews/issues/10400)
