### 2026-04-18 kieker-monitoring/moobench#52 @shrishar
[JOSS Review] Add core functionality of the software in detail
System design in the paper is not very clear. Some of the figures are missing. README document has some design but the paper lacks proper design and architecture flow. Please update the paper.
https://github.com/openjournals/joss-reviews/issues/10400
