### 2026-04-18 kieker-monitoring/moobench#53 @shrishar
[JOSS Review] Add some test cases
https://github.com/openjournals/joss-reviews/issues/10400
