### 2026-04-19 c461ae7
Add correctness tests instead of GH actions for every framework
