+++ .github/workflows/testCorrectness.yaml
@@ -1,0 +1,39 @@
+---
+name: Execution Check Kieker-java
+on:
+  push:
+  pull_request:
+  workflow_dispatch:
+  schedule:
+    - cron: 0 */8 * * *
+
+jobs:
+  build:
+    strategy:
+      matrix:
+        os:
+          - ubuntu-latest
+        java:
+          - 17
+          - 21
+      fail-fast: false
+    runs-on: ${{ matrix.os }}
+    name: Java ${{ matrix.java }} OS ${{ matrix.os }} sample
+    timeout-minutes: 15 # For check of functional runnability, 15 minutes should be more than sufficient, otherwise, developer should be notified
+    steps:
+      - uses: actions/checkout@v4
+      - uses: r-lib/actions/setup-r@v2
+        with:
+          r-version: '3.5.3'
+      - name: Set up JDK ${{ matrix.java }}
+        uses: actions/setup-java@v4
+        with:
+          java-version: ${{ matrix.java }}
+          distribution: 'temurin'
+      - name: Setup (Compile and move benchmark application)
+        ./setup.sh
+      - name: Kieker correctness check
+        run: |
+          cd frameworks/Kieker-java/testing
+          ../../../test/bats/bin/bats --print-output-on-failure .
+
