### 2026-04-19 f0ada92
Rename workflow and fix syntax
