### 2026-04-19 c990754
Test result file creation for OpenTelemetry
