### 2026-04-19 cd8497d
Add zipkin trace creation test
