### 2026-04-19 5baca0c
Extract testkiekerTraceCreation.sh
