### 2026-04-19 99d22e1
Add calltree.pdf and hint in paper to fix GH generated paper
