### 2026-04-19 1a84a6d
Add calltree.pdf and hint in paper to fix GH generated paper
