### 2026-04-19 e1a2470
Add architecture graph
