+++ .github/workflows/testCorrectness.yaml
@@ -22,6 +22,9 @@ jobs:
     timeout-minutes: 15 # For check of functional runnability, 15 minutes should be more than sufficient, otherwise, developer should be notified
     steps:
       - uses: actions/checkout@v4
+        with:
+          submodules: true # necessary for bats
+          fetch-depth: 0
       - uses: r-lib/actions/setup-r@v2
         with:
           r-version: '3.5.3'
