### 2026-04-19 30d4a04
Remove unnecessary workflow - compilation check is done with tests
