+++ .github/workflows/buildBenchmark.yaml
@@ -1,26 +1,0 @@
-# This workflow will build a Java project with Maven
-# For more information see: https://help.github.com/actions/language-and-framework-guides/building-and-testing-java-with-maven
-
-name: Build MooBench for Compilation Check
-
-on: [push, pull_request, workflow_dispatch]
-
-jobs:
-  build:
-    strategy:
-        matrix:
-          os: [ubuntu-latest]
-          java: [ 17, 21 ]
-        fail-fast: false
-    runs-on: ${{ matrix.os }}
-    name: Java ${{ matrix.java }} OS ${{ matrix.os }} sample
-    timeout-minutes: 15 # For check of functional runnability, 15 minutes should be more than sufficient, otherwise, developer should be notified
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up JDK ${{ matrix.java }}
-      uses: actions/setup-java@v4
-      with:
-        java-version: ${{ matrix.java }}
-        distribution: 'temurin'
-    - name: Assemble Project
-      run: ./gradlew assemble
