### 2026-04-19 2a84483
Remove execution check Kieker-java and OpenTelemetry-java, since their contents have been moved to testCorrectness.yaml
