+++ .github/workflows/executeOpenTelemetryJava.yaml
@@ -1,47 +1,0 @@
----
-name: Execution Check OpenTelemetry-java
-on:
-  - push
-  - pull_request
-  - workflow_dispatch
-jobs:
-  build:
-    strategy:
-      matrix:
-        os:
-          - ubuntu-latest
-        java:
-          - 17
-          - 21
-      fail-fast: false
-    runs-on: ${{ matrix.os }}
-    name: Java ${{ matrix.java }} OS ${{ matrix.os }} sample
-    timeout-minutes: 15 # For check of functional runnability, 15 minutes should be more than sufficient, otherwise, developer should be notified
-    steps:
-      - uses: actions/checkout@v4
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: '3.5.3'
-      - name: Set up JDK ${{ matrix.java }}
-        uses: actions/setup-java@v4
-        with:
-          java-version: ${{ matrix.java }}
-          distribution: 'temurin'
-      - name: Run minimal parameter experiment and check success
-        run: |
-          ./setup.sh
-          cd frameworks/OpenTelemetry-java/ 
-          export SLEEP_TIME=0
-          export NUM_OF_LOOPS=1 
-          export TOTAL_NUM_OF_CALLS=10 
-          export RECURSION_DEPTH=10 
-          export MOOBENCH_CONFIGURATIONS="0 1 2 3 4 6"
-          ./benchmark.sh 
-          unzip results-OpenTelemetry-java/results.zip 
-          measuredValues=$(cat raw-1-10-* | wc -l) 
-          if [ $measuredValues -ne 60 ] 
-          then
-            echo "It should be 60 measured values, but was $measuredValues"
-            exit 1
-          fi
-
