+++ .github/workflows/testCorrectness.yaml
@@ -21,7 +21,7 @@ jobs:
     name: Java ${{ matrix.java }} OS ${{ matrix.os }} sample
     timeout-minutes: 15 # For check of functional runnability, 15 minutes should be more than sufficient, otherwise, developer should be notified
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
         with:
           submodules: true # necessary for bats
           fetch-depth: 0
