+++ .github/workflows/testCorrectness.yaml
@@ -29,7 +29,7 @@ jobs:
         with:
           r-version: '3.5.3'
       - name: Set up JDK ${{ matrix.java }}
-        uses: actions/setup-java@v4
+        uses: actions/setup-java@v5
         with:
           java-version: ${{ matrix.java }}
           distribution: 'temurin'
