### 2026-04-19 3b198f2
Update architecture graph draft
