+++ moobench-paper/architecture.tex
@@ -12,33 +12,37 @@
       
       \node[component, right=0.15cm of benchmarkJ.west, yshift=0.25cm] (benchmark) {benchmark};
       \node[component, below=0.2cm of benchmark] (application) {application};
+      
+      \draw [->, thick] ([xshift=-0.5cm]benchmark.south) -- ([xshift=-0.5cm]application.north);
+      \draw [->, thick] ([xshift=0.5cm]application.north) -- ([xshift=0.5cm]benchmark.south);
 
       \node[shape=rectangle, minimum width=2.25cm, below=0.5cm of benchmarkJ] (pybenchmark) {pybenchmark\\[1.25cm]};
       \node[component, right=0.15cm of pybenchmark.west, yshift=0.25cm] (benchmark2) {benchmark};
       \node[component, below=0.2cm of benchmark2] (application2) {application};
+      
+      \draw [->, thick] ([xshift=-0.5cm]benchmark2.south) -- ([xshift=-0.5cm]application2.north);
+      \draw [->, thick] ([xshift=0.5cm]application2.north) -- ([xshift=0.5cm]benchmark2.south);
 
-      \node (OpenTelemetry-java) [framework, right=1.0cm of benchmarkJ.north east, anchor=north west] {OpenTelemetry-java\\[1.25cm]};
-      \node (Kieker-java) [framework, below=0.1cm of OpenTelemetry-java] {Kieker-java};
-      \node (other) [framework, below=0.1cm of Kieker-java] {...};
-      \node (OpenTelemetry-python) [framework, below=of other] {OpenTelemetry-python};
+      \node (OpenTelemetry-java) [framework, right=1.0cm of benchmarkJ.north east, anchor=north west] {OpenTelemetry-java\\[1.1cm]};
+      \node (other) [framework, below=0.1cm of OpenTelemetry-java] {...};
+      \node (OpenTelemetry-python) [framework, below=of other] {OpenTelemetry-python\\[1.1cm]};
 
       \node[above=of benchmarkJ.north, anchor=north, align=center, rounded corners, minimum width=3.0cm, line width=1pt] (tools) {\textbf{tools}\\[4.5cm]};
       \node[above=of OpenTelemetry-java).north east, xshift=0.0cm, anchor=north, align=center, rounded corners, minimum width=5.5cm, line width=1pt] (tools) {\textbf{frameworks}\\[4.5cm]};
-      %\draw[draw, rounded corners, line width=1pt] ($(Kieker-java-aspectj.north west) - (0.15, -0.5cm)$) rectangle ($(Kieker-java-aspectj.north east) - (-2.75, 1.5cm)$) node [draw=none,pos=.2,xshift=-1.25cm] {\textbf{frameworks}};
-
-      \draw [->, thick] (benchmarkJ) -- (Kieker-java.west);
-      \draw [->, thick] (benchmarkJ) -- (OpenTelemetry-java.west);
-      \draw [->, thick] (benchmarkJ) -- (other.west);
-      \draw [->, thick] (pybenchmark) -- (OpenTelemetry-python.west);
       
       \node[component, below=of OpenTelemetry-java.north, xshift=-1cm] (benchmarkSH) {benchmark.sh};
       \node[component, right=of benchmarkSH] (functionsSH) {functions.sh};
-      \node[component, dotted, below=0.2cm of benchmarkSH]  (applicationInstrumented) {application};
-      
+      \node[component, dotted, below=0.2cm of benchmarkSH]  (applicationInstrumented) {application};      
       \draw [->, thick] (benchmarkSH) -- (functionsSH);
-      \draw [->, thick] (application) -- (applicationInstrumented);
-      
[…]
