+++ .github/workflows/benchmarkScouterJava.yaml
@@ -14,7 +14,7 @@ jobs:
     name: Performance measurement
     runs-on: ubuntu-24.04
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
       - uses: r-lib/actions/setup-r@v2
         with:
           r-version: '3.5.3'
@@ -22,7 +22,7 @@ jobs:
         with:
           go-version: "stable"
       - name: Set up JDK ${{ matrix.java }}
-        uses: actions/setup-java@v4
+        uses: actions/setup-java@v5
         with:
           distribution: 'temurin'
           java-version: 21
