### 2026-04-20 b0216f3
Correctness Test for the inspectIT-java framework
