### 2026-04-20 a995f8d
Correctness Test for the inspectIT-java framework
