### 2026-04-20 5a51fe7
Added re_try counts for retrieving spans
