### 2026-04-20 kieker-monitoring/moobench#53 @DaGeRe
Thanks for the hint.
We agree that the old approach via bash scripts in GH actions is not sufficient. Therefore, we started building test scripts using bats for all frameworks, looking like this: https://github.com/kieker-monitoring/moobench/tree/main/frameworks/Kieker-java/tests I think it will take us roughly one week to have this for all frameworks.

These test cases are executed using: https://github.com/kieker-monitoring/moobench/blob/main/.github/workflows/testCorrectness.yaml

I will notify you here again when this is finished.
