### 2026-04-20 kieker-monitoring/moobench#52 @DaGeRe
Thanks for the hint, the figure required compiling it locally, which is fixed now (https://github.com/kieker-monitoring/moobench/commit/99d22e1268fc8595099b73bb722b0955d0c9bcf9).
I'll update the paper within the next week, adding more details to the architecture description.
