### 2026-04-20 cf8db99
Shutdown zipkin process before exiting
