### 2026-04-21 086bb42
Rename tools/benchmark -> SuT-java, tools/pybenchmark -> SuT-python
