+++ .github/workflows/benchmarkKiekerJava2.yaml
@@ -42,7 +42,7 @@ jobs:
           export THREADS=2
           export NUM_OF_LOOPS=10 
           export MOOBENCH_CONFIGURATIONS="0 1 2 4"
-          ./benchmark.sh 
+          ./start.sh 
           cd results-Kieker-java
           unzip results.zip 
           ../../../analysis/getGHActionsFormat.sh &> output.json
