+++ .github/workflows/benchmarkKiekerJava8.yaml
@@ -33,7 +33,7 @@ jobs:
           export THREADS=8
           export NUM_OF_LOOPS=10 
           export MOOBENCH_CONFIGURATIONS="0 1 2 4"
-          ./benchmark.sh 
+          ./start.sh 
           cd results-Kieker-java
           unzip results.zip 
           ../../../analysis/getGHActionsFormat.sh &> output.json
