+++ .github/workflows/benchmarkKiekerJava.yaml
@@ -43,7 +43,7 @@ jobs:
           cd frameworks/Kieker-java/
           export NUM_OF_LOOPS=10 
           export MOOBENCH_CONFIGURATIONS="0 1 2 4"
-          ./start.sh 
+          ./measure.sh 
           cd results-Kieker-java
           unzip results.zip 
           ../../../analysis/getGHActionsFormat.sh &> output.json
