### 2026-04-21 8c11d91
More concise script names: measure instead of benchmark
