+++ .github/workflows/benchmarkScouterJava.yaml
@@ -41,7 +41,7 @@ jobs:
           cd frameworks/Scouter-java/
           export MOOBENCH_CONFIGURATIONS="0 1 2"
           export NUM_OF_LOOPS=10
-          ./benchmark.sh 
+          ./measure.sh 
           cd results-Scouter-java
           unzip results.zip
           pwd
