+++ .github/workflows/benchmarkElasticAPM.yaml
@@ -43,7 +43,7 @@ jobs:
           cd frameworks/elasticapm-java/
           export NUM_OF_LOOPS=10 
           export MOOBENCH_CONFIGURATIONS="0 1 2 3"
-          ./benchmark.sh 
+          ./measure.sh 
           cd results-elasticapm-java
           unzip results.zip 
           ../../../analysis/getGHActionsFormat.sh &> output.json
