### 2026-04-21 703fadf
Increase timeout to 2 hours, since testing multiple frameworks takes longer
