### 2026-04-21 cd26826
Remove legacy script
