### 2026-04-21 82f6fcc
Fix OpenTelemetry environment variable
