### 2026-04-23 cbef601
Skywalking APM 10.3.0 will fail due to Java25
