### 2026-04-27 437516b
Add basic bats test for Skywalking
