### 2026-04-27 dc3a713
Add basic test for scouter
