### 2026-04-28 fecd160
Remove inspectIT test, since its already in testCorrectness
