### 2026-04-28 fcc8a44
Add Scouter-java and Skywalking-java to the unit tests
