### 2026-04-30 c0ddce1
Remove outdated execute workflows for Scouter and Skywalking
