+++ .github/workflows/executeScouterJava.yaml
@@ -1,49 +1,0 @@
----
-name: Execution Check Scouter-java
-on:
-  push:
-  pull_request:
-  workflow_dispatch:
-  schedule:
-    - cron: 0 */8 * * *
-
-jobs:
-  build:
-    strategy:
-      matrix:
-        os:
-          - ubuntu-latest
-        java:
-          - 17
-          - 21
-      fail-fast: false
-    runs-on: ${{ matrix.os }}
-    name: Java ${{ matrix.java }} OS ${{ matrix.os }} sample
-    timeout-minutes: 15 # For check of functional runnability, 15 minutes should be more than sufficient, otherwise, developer should be notified
-    steps:
-      - uses: actions/checkout@v4
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: '3.5.3'
-      - name: Set up JDK ${{ matrix.java }}
-        uses: actions/setup-java@v4
-        with:
-          java-version: ${{ matrix.java }}
-          distribution: 'temurin'
-      - name: Run minimal parameter experiment and check success
-        run: |
-          ./setup.sh
-          cd frameworks/Scouter-java/ 
-          export SLEEP_TIME=0
-          export NUM_OF_LOOPS=1 
-          export TOTAL_NUM_OF_CALLS=10 
-          export RECURSION_DEPTH=10 
-          ./measure.sh 
-          unzip results-Scouter-java/results.zip 
-          measuredValues=$(cat raw-1-10-* | wc -l) 
-          if [ $measuredValues -ne 30 ] 
-          then
-            echo "It should be 30 measured values, but was $measuredValues"
-            exit 1
-          fi
-
