### 2026-04-30 2959c18
Remove outdated execute workflows for Scouter and Skywalking
