+++ .github/workflows/executeSkywalkingJava.yaml
@@ -1,49 +1,0 @@
----
-name: Execution Check Skywalking-java
-on:
-  push:
-  pull_request:
-  workflow_dispatch:
-  schedule:
-    - cron: 0 */8 * * *
-
-jobs:
-  build:
-    strategy:
-      matrix:
-        os:
-          - ubuntu-latest
-        java:
-          - 17
-          - 21
-      fail-fast: false
-    runs-on: ${{ matrix.os }}
-    name: Java ${{ matrix.java }} OS ${{ matrix.os }} sample
-    timeout-minutes: 20
-    steps:
-      - uses: actions/checkout@v4
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: '3.5.3'
-      - name: Set up JDK ${{ matrix.java }}
-        uses: actions/setup-java@v4
-        with:
-          java-version: ${{ matrix.java }}
-          distribution: 'temurin'
-      - name: Run minimal parameter experiment and check success
-        run: |
-          ./setup.sh
-          cd frameworks/Skywalking-java/ 
-          export SLEEP_TIME=0
-          export NUM_OF_LOOPS=1 
-          export TOTAL_NUM_OF_CALLS=10 
-          export RECURSION_DEPTH=10 
-          ./measure.sh 
-          unzip results-Skywalking-java/results.zip 
-          measuredValues=$(cat raw-1-10-* | wc -l) 
-          if [ $measuredValues -ne 40 ] 
-          then
-            echo "It should be 40 measured values, but was $measuredValues"
-            exit 1
-          fi
-
