### 2026-04-30 795b51c
Check kieker-python with bats
