+++ .github/workflows/testCorrectness.yaml
@@ -56,3 +56,7 @@ jobs:
         run: |
           cd frameworks/Skywalking-java/tests
           ../../../test/bats/bin/bats --print-output-on-failure *.sh
+      - name: Skywalking correctness check
+        run: |
+          cd frameworks/Kieker-python/tests
+          ../../../test/bats/bin/bats --print-output-on-failure *.sh
