### 2026-04-30 35285ab
Add pinpoint-java test
