### 2026-04-30 0135aff
Add pinpoint-java test
