### 2026-04-30 b933142
Add architecture overview and adapt to changed architecture
