+++ moobench-paper/architecture.tex
@@ -8,16 +8,16 @@
   framework/.style={draw, minimum width=5.0cm, shape=rectangle},
   component/.style={draw, minimum width=2.0cm, shape=rectangle}]
 
-      \node[shape=rectangle, minimum width=2.25cm] (benchmarkJ) {benchmark\\[1.25cm]};
+      \node[shape=rectangle, minimum width=2.25cm] (benchmarkJ) {SuT-java\\[1.25cm]};
       
-      \node[component, right=0.15cm of benchmarkJ.west, yshift=0.25cm] (benchmark) {benchmark};
+      \node[component, right=0.15cm of benchmarkJ.west, yshift=0.25cm] (benchmark) {measurement\\Harness};
       \node[component, below=0.2cm of benchmark] (application) {application};
       
       \draw [->, thick] ([xshift=-0.5cm]benchmark.south) -- ([xshift=-0.5cm]application.north);
       \draw [->, thick] ([xshift=0.5cm]application.north) -- ([xshift=0.5cm]benchmark.south);
 
-      \node[shape=rectangle, minimum width=2.25cm, below=0.5cm of benchmarkJ] (pybenchmark) {pybenchmark\\[1.25cm]};
-      \node[component, right=0.15cm of pybenchmark.west, yshift=0.25cm] (benchmark2) {benchmark};
+      \node[shape=rectangle, minimum width=2.25cm, below=0.5cm of benchmarkJ] (pybenchmark) {SuT-python\\[1.25cm]};
+      \node[component, right=0.15cm of pybenchmark.west, yshift=0.25cm] (benchmark2) {measurement\\Harness};
       \node[component, below=0.2cm of benchmark2] (application2) {application};
       
       \draw [->, thick] ([xshift=-0.5cm]benchmark2.south) -- ([xshift=-0.5cm]application2.north);
@@ -30,14 +30,14 @@
       \node[above=of benchmarkJ.north, anchor=north, align=center, rounded corners, minimum width=3.0cm, line width=1pt] (tools) {\textbf{tools}\\[4.5cm]};
       \node[above=of OpenTelemetry-java).north east, xshift=0.0cm, anchor=north, align=center, rounded corners, minimum width=5.5cm, line width=1pt] (tools) {\textbf{frameworks}\\[4.5cm]};
       
-      \node[component, below=of OpenTelemetry-java.north, xshift=-1cm] (benchmarkSH) {benchmark.sh};
+      \node[component, below=of OpenTelemetry-java.north, xshift=-1cm] (benchmarkSH) {measure.sh};
       \node[component, right=of benchmarkSH] (functionsSH) {functions.sh};
       \node[component, dotted, below=0.2cm of benchmarkSH]  (applicationInstrumented) {application};      
       \draw [->, thick] (benchmarkSH) -- (functionsSH);
       \draw [->, thick] (benchmarkJ) -- (applicationInstrumented);
       \draw (functionsSH) edge [bend left, ->, thick] node [below, draw=none, xshift=0.5cm] {\textit{Instrumentation}} (applicationInstrumented.east);
       
-      \node[component, below=of OpenTelemetry-python.north, xshift=-1cm] (benchmarkSH2) {benchmark.sh};
+      \node[component, below=of OpenTelemetry-python.north, xshift=-1cm] (benchmarkSH2) {measure.sh};
       \node[component, right=of benchmarkSH2] (functionsSH2) {functions.sh};
       \node[component, dotted, below=0.2cm of benchmarkSH2]  (applicationInstrumented2) {application};      
       \draw [->, thick] (benchmarkSH2) -- (functionsSH2);
