### 2026-04-30 kieker-monitoring/moobench#52 @DaGeRe
After we had a look at the architecture, we noticed that several changes were necessary to clarify terminology (many things were called "benchmark" that had a different purpose). We completed this refactoring and added an architecture overview graph. Could you have a look again?
