### 2026-04-30 21fb1ac
Adapt to changed architecture
