### 2026-04-30 kieker-monitoring/moobench#51 @DaGeRe
I've now added an example of MooBench results: https://github.com/kieker-monitoring/moobench#benchmark-execution-1

The real world usage of MooBench is using it for measuring observability overhead. The results are these statistics (which then can be used for further analysis). Does this make it clearer?
