### 2026-04-30 48edbe9
Remove MOOBENCH_CONFIGURATIONS so all default configurations are used
