### 2026-05-03 #10400 @jatkinson1000
Thank you @shrishar  - @DaGeRe please look at addressing these and keep us updated via the issues.

@abhishektiwari and @Ahanaf-Aziz I see you have both worked through the early parts of the review checklist.
Might you be able to look at these over the next week please as JOSS likes to aim for a 4-week turnaround for reviews, at least the first pass to provide authors with issues they need to work on?

If, for any reason, you will need more time please just let us know so that we and the authors are aware of the situation.
