### 2026-06-12 53c9f20
Pinpoint does not work with Java 25 yet, so remove Java 25 for now
