### 2026-06-16 3c75d9c
Added RSECon26 logo
See https://github.com/openjournals/joss-reviews/issues/10400#issuecomment-4716788438
