### 2026-06-16 58fba20
Added RSECon26 logo
See https://github.com/openjournals/joss-reviews/issues/10400#issuecomment-4716788438
