### 2026-06-26 64478b9
Editorial corrections from JOSS review
