+++ moobench-paper/paper.md
@@ -41,13 +41,13 @@ The MooBench benchmark was originally developed to examine the performance overh
 
 # Statement of need
 
-Observability tools are crucial for managing software system health and optimizing operational costs; consequently, a wide variety of tools compete in the market [@GartnerObservability2025]. The relevance of distributed tracing tools increased with the rise of microservice architectures, where system behavior across microservices needs to be understood [@sigelman2010dapper]. While observability generally consists of the three pillars traces, metrics, and logs, especially distributed tracing is gaining wide adoption [@janes2023open]. Concerning the categorization of research software [@Categorization2025], MooBench is technology research software [@TRS2026]: It is developed and used in research, but it is also intended for use in non-academic contexts for evaluating and selecting observability tools.
+Observability tools are crucial for managing software system health and optimizing operational costs; consequently, a wide variety of tools compete in the market [@GartnerObservability2025]. The relevance of distributed tracing tools increased with the rise of microservice architectures, where system behavior across microservices needs to be understood [@sigelman2010dapper]. While observability generally consists of the three pillars: traces, metrics, and logs, with distributed tracing especially gaining wide adoption [@janes2023open]. Concerning the categorization of research software [@Categorization2025], MooBench is technology research software [@TRS2026]: It is developed and used in research, but it is also intended for use in non-academic contexts for evaluating and selecting observability tools.
 
 # State of the field
 
 In the context of microservices, various studies have examined the overhead of tracing. For example, @nou2025investigating investigated the overhead of OpenTelemetry and Elastic APM using open-source applications, and examined the root causes of the overhead using profiling. @ahmed2016studying compared the overhead of commercial observability tools and the open-source tool Pinpoint for regression detection. @eder2023comparison compared the overhead using serverless applications.
 
-Besides the microservice context, tracing overhead has been examined for operating systems and High-Performance Computing (HPC). For operating systems, the overhead of tools for tracing the kernel itself [@desnoyers2006lttng; @gebai2018tracers] and eBPF, that allows to trace both within the kernel and the user space [@volpertEBPFOverhead; @domaschka2023using], has been examined in case studies. In the context of HPC, research has addressed the impact of overhead [@hunold2022overhead] and the combination of instrumentation and sampling to reduce it [@ilsche2015combining].
+Besides the microservice context, tracing overhead has been examined for operating systems and High-Performance Computing (HPC). For operating systems, the overhead of tools for tracing the kernel itself [@desnoyers2006lttng; @gebai2018tracers] and eBPF, that allows tracing within both the kernel and the user space [@volpertEBPFOverhead; @domaschka2023using], have been examined in case studies. In the context of HPC, research has addressed the impact of overhead [@hunold2022overhead] and the combination of instrumentation and sampling to reduce it [@ilsche2015combining].
 
 The aforementioned studies utilize macrobenchmarks to examine overhead in realistic use cases. While macrobenchmarks provide an indication of overhead in similar scenarios, they often fail to isolate the baseline overhead or identify its specific sources. When looking at specific distributed cases, the overhead sources are influenced by the processes running in parallel, their scheduling, and the network behavior. MooBench addresses this by providing a microbenchmark that measures the fundamental overhead of observability frameworks in a controlled environment. Using different configurations, it enables factorial experiments that isolate root causes of overhead, specifically distinguishing between instrumentation, data collection, and data serialization.
 
[…]
