### 2026-06-26 956111a
Fix formatting of title in bibliography
