+++ moobench-paper/paper.bib
@@ -147,7 +147,7 @@ series = {ICPE '25}
 }
 
 @Article{georges2007statistically,
-  title	= {Statistically Rigorous Java Performance Evaluation},
+  title	= {Statistically Rigorous {Java} Performance Evaluation},
   author	= {Georges, Andy and Buytaert, Dries and Eeckhout, Lieven},
   journal	= {ACM SIGPLAN Notices},
   volume	= {42},
