### 2026-06-26 07f0a47
Clarification of ambiguous text on re-reading.
I think direct source was meant to be a separate item upon re-reading.
