+++ moobench-paper/paper.md
@@ -107,7 +107,7 @@ In our most recent study [@reichelt2026benchmarking], we integrated Pinpoint, Sc
 
 @reichelt2023towards examined different options for overhead reduction for tracing: Using source code instrumentation instead of AspectJ [@eclipse2026aspectj], storing only limited metadata, using a different queue and aggregating performance data before writing it to the monitoring queue. Using MooBench, they found that on their examined hardware, they could reduce the overhead from 4.77 ns to 0.4 ns per method call.
 
-@reichelt2024overhead compared the overhead of the instrumentation frameworks AspectJ, ByteBuddy [@rafael2025bytebuddy], DiSL [@marek2012disl], and Javassist [@shigeru2023disl] direct source code instrumentation. To do so, they extended the Kieker-java scripts of MooBench. Through these extensions of Kieker and MooBench, they found that while source code instrumentation has the lowest overhead, ByteBuddy, DiSL and Javassist also have comparably low overhead. AspectJ causes significantly higher overhead than the others. Based on these findings, the Kieker framework transitioned to using the Kieker ByteBuddy agent as its default.
+@reichelt2024overhead compared the overhead of the instrumentation frameworks AspectJ, ByteBuddy [@rafael2025bytebuddy], DiSL [@marek2012disl], Javassist [@shigeru2023disl], and direct source code instrumentation. To do so, they extended the Kieker-java scripts of MooBench. Through these extensions of Kieker and MooBench, they found that while source code instrumentation has the lowest overhead, ByteBuddy, DiSL and Javassist also have comparably low overhead. AspectJ causes significantly higher overhead than the others. Based on these findings, the Kieker framework transitioned to using the Kieker ByteBuddy agent as its default.
 
 ## Continuous execution for regression benchmarking
 
