+++ .github/workflows/testCorrectness.yaml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         os:
-          - ubuntu-latest
+          - ubuntu-26.04
         java:
           - 17
           - 21
