+++ .github/workflows/testCorrectness.yaml
@@ -27,7 +27,7 @@ jobs:
           fetch-depth: 0
       - uses: r-lib/actions/setup-r@v2
         with:
-          r-version: '3.5.3'
+          r-version: '4.4.2'
       - name: Set up JDK ${{ matrix.java }}
         uses: actions/setup-java@v5
         with:
