+++ .github/workflows/executeKiekerPython.yaml
@@ -20,7 +20,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: r-lib/actions/setup-r@v2
         with:
-          r-version: '3.5.3'
+          r-version: '4.4.2'
       - name: Set up JDK ${{ matrix.java }}
         uses: actions/setup-java@v4
         with:
