+++ .github/workflows/benchmarkKiekerJava12.yaml
@@ -19,7 +19,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: r-lib/actions/setup-r@v2
         with:
-          r-version: '3.5.3'
+          r-version: '4.4.2'
       - uses: actions/setup-go@v4
         with:
           go-version: "stable"
