### 2026-06-30 7467687
Remove go - was only present since it was part of the benchmark-action demo
