+++ .github/workflows/benchmarkKiekerJava12.yaml
@@ -20,9 +20,6 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
         with:
           r-version: '4.4.2'
-      - uses: actions/setup-go@v4
-        with:
-          go-version: "stable"
       - name: Set up JDK ${{ matrix.java }}
         uses: actions/setup-java@v4
         with:
