### 2026-06-30 7494dc4
Rephrase software design introduction
