### 2026-06-30 e9d2dec
Rephrase again
