### 2026-07-06 dddb04a
Fix javassist reference
