+++ moobench-paper/paper.bib
@@ -427,12 +427,13 @@ author = "Wilhelm Hasselbring and André {van Hoorn}"
   series = {AOSD '12}
 }
 
-@online{shigeru2023dissle,
-  title={Javassist},
-  organization={Shigeru Chiba},
-  date={2023-12-17T02:19:53+01:00},
-  urldate={2026-03-24T12:40:38+01:00},
-  url={https://www.javassist.org/},
+@inproceedings{shigeru2003javassist,
+  title={An easy-to-use toolkit for efficient Java bytecode translators},
+  author={Chiba, Shigeru and Nishizawa, Muga},
+  booktitle={Proc.\ Int.\ Conf.\ on Generative Programming and Component Engineering},
+  pages={364--376},
+  year={2003},
+  organization={Springer}
 }
 
 @article{Categorization2025,
