### 2026-07-06 2ba7da6
Add factorial experiments for timers
