+++ frameworks/Kieker-java/labels.sh
@@ -5,3 +5,5 @@ TITLE[3]="Text file"
 TITLE[4]="Binary file"
 TITLE[5]="Binary TCP"
 TITLE[6]="Binary file (Milli timer)"
+TITLE[7]="Binary file (Zero timer)"
+TITLE[8]="Binary TCP (Zero timer)"
