### 2026-07-06 b0e1076
Add doi for javassist reference
