+++ moobench-paper/paper.bib
@@ -433,7 +433,8 @@ author = "Wilhelm Hasselbring and André {van Hoorn}"
   booktitle={Proc.\ Int.\ Conf.\ on Generative Programming and Component Engineering},
   pages={364--376},
   year={2003},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-540-39815-8_22}
 }
 
 @article{Categorization2025,
