### 2026-07-06 fa04477
Fix duplicated reference name
