+++ moobench-paper/paper.bib
@@ -317,7 +317,7 @@ series = {ICPE '25}
   url={https://dl.gi.de/handle/20.500.12116/47944}
 }
 
-@article{reichelt2024overhead,
+@article{reichelt2024overhead2,
   title={Overhead Measurement Noise in Different Runtime Environments},
   author={Reichelt, David Georg and Jung, Reiner and van Hoorn, Andr{\'e}},
   journal={Softwaretechnik-Trends (Proceedings of the 15th Symposium on Software Performance 2024)},
