+++ moobench-paper/paper.md
@@ -111,7 +111,7 @@ In our most recent study [@reichelt2026benchmarking], we integrated Pinpoint, Sc
 
 ## Continuous execution for regression benchmarking
 
-Several authors worked on continuously executing MooBench benchmarks for spotting performance regressions in observability frameworks. @waller2014application started to include MooBench's benchmarking into CI. To reduce perturbations from concurrently running software, they configured Jenkins to execute measurements on a separate server. This setup has been running continuously to date, with regular hardware and system software updates. Additionally, a performance measurement setup for GitHub Actions was developed [@reichelt2024overhead]. This additional setup utilizes GitHub's default runners, that, despite sharing the execution infrastructure with others, show low standard deviation and lower execution times than the original Jenkins setup. The measurement data from GitHub have been uploaded to Nyrkiö, a performance change detection tool for GitHub Actions. Using this tool, past regressions could be reproduced [@yang2025detection]. In the future, it is planned to automatically detect performance regressions.
+Several authors worked on continuously executing MooBench benchmarks for spotting performance regressions in observability frameworks. @waller2014application started to include MooBench's benchmarking into CI. To reduce perturbations from concurrently running software, they configured Jenkins to execute measurements on a separate server. This setup has been running continuously to date, with regular hardware and system software updates. Additionally, a performance measurement setup for GitHub Actions was developed [@reichelt2024overhead2]. This additional setup utilizes GitHub's default runners, that, despite sharing the execution infrastructure with others, show low standard deviation and lower execution times than the original Jenkins setup. The measurement data from GitHub have been uploaded to Nyrkiö, a performance change detection tool for GitHub Actions. Using this tool, past regressions could be reproduced [@yang2025detection]. In the future, it is planned to automatically detect performance regressions.
 
 # AI usage disclosure
 
