+++ frameworks/pinpoint-java/functions.sh
@@ -149,7 +149,7 @@ function waitForStartup {
   fi
 }
 
-export PINOT_VERSION=1.5.0
+export PINOT_VERSION=1.5.1
 function startPinot() {
    echo "Starting Pinot $PINOT_VERSION"
    if [ ! -d apache-pinot-$PINOT_VERSION-bin ]; then
