### 2026-07-07 077b5cb
minor changes in paper.md
