### 2026-07-07 5f6d01a
minor changes in paper.bib
