+++ moobench-paper/paper.bib
@@ -38,7 +38,7 @@
 
 @article{gebai2018tracers,
 author     = {Gebai, Mohamad and Dagenais, Michel R.},
-title      = {Survey and Analysis of Kernel and Userspace Tracers on Linux: Design, Implementation, and Overhead},
+title      = {Survey and Analysis of Kernel and Userspace Tracers on {L}inux: Design, Implementation, and Overhead},
 year       = {2018},
 issue_date = {March 2019},
 publisher  = {Association for Computing Machinery},
@@ -241,7 +241,7 @@ series = {ICPE '25}
 
 
 @inproceedings{knoche2018using,
-  title={{Using the Raspberry Pi and Docker for Replicable Performance Experiments: Experience Paper}},
+  title={Using the {R}aspberry {P}i and {D}ocker for Replicable Performance Experiments: Experience Paper},
   author={Knoche, Holger and Eichelberger, Holger},
   booktitle={Proc.\ 8th ACM/SPEC Int.\ Conf.\ on Performance Engineering},
   pages={305--316},
@@ -308,7 +308,7 @@ series = {ICPE '25}
 }
 
 @article{yang2025detection,
-  title={{Detection of Performance Changes in MooBench Results Using Nyrkiö on GitHub Actions}},
+  title={Detection of Performance Changes in {MooBench} Results Using {N}yrkiö on {GitHub} {A}ctions},
   author={Yang, Shinhyung and Reichelt, David Georg and Ingo, Henrik and Hasselbring, Wilhelm},
   journal={Softwaretechnik-Trends (Proceedings des 16th Symposium on Software Performance 2025)},
   volume=45,
@@ -383,7 +383,7 @@ author = "Wilhelm Hasselbring and André {van Hoorn}"
 @inproceedings{yang2025kieker,
   author={Yang, Shinhyung and Reichelt, David Georg and Jung, Reiner and Hansson, Marcel and Hasselbring, Wilhelm},
   booktitle={Companion of the 16th ACM/SPEC International Conference on Performance Engineering},
-  title = 	 {{The Kieker Observability Framework Version 2}},
+  title = 	 {The {K}ieker Observability Framework Version 2},
   booktitle = {Companion of the 16th ACM/SPEC International Conference on Performance Engineering (ICPE 2025)},
   year = 	 2025,
   DOI = {10.1145/3680256.3721972},
@@ -428,7 +428,7 @@ author = "Wilhelm Hasselbring and André {van Hoorn}"
 }
 
 @inproceedings{shigeru2003javassist,
-  title={An easy-to-use toolkit for efficient Java bytecode translators},
+  title={An easy-to-use toolkit for efficient {J}ava bytecode translators},
   author={Chiba, Shigeru and Nishizawa, Muga},
   booktitle={Proc.\ Int.\ Conf.\ on Generative Programming and Component Engineering},
   pages={364--376},
