### 2026-07-08 711aa12
Add toolpaper CITATION.cff
