### 2026-07-09 667d497
Also test correctness for Java 25
