### 2026-07-09 9a7123a
Skip pinpoint if Java 25 is active; fix can be added later
