### 2026-07-09 598bce3
Completely disable unit tests for pinpoint Java 25
