### 2026-07-09 0715a0a
Go back to Java 17 for benchmarking pinpoint
