+++ .github/workflows/benchmarkPinpointJava.yaml
@@ -24,7 +24,7 @@ jobs:
         uses: actions/setup-java@v5
         with:
           distribution: 'temurin'
-          java-version: 21
+          java-version: 17
       - name: Start ssh-agent and add private key
         uses: webfactory/ssh-agent@v0.10.0
         with:
