### 2026-07-11 5e4b446
Leave out Java 21
