### 2026-07-24 6d5fbc5
Kieker-python fixes
1. Added new hook routines for the post import hook.
2. Minor updates has been made as kieker-lang-pack-python 0.0.2
