### 2026-07-31 c5a65de
Fix MOOBENCH_CONFIGURATIONS
