+++ frameworks/OpenTelemetry-python/labels.sh
@@ -1,4 +1,3 @@
-MOOBENCH_CONFIGURATIONS="0 1 2"
 TITLE[0]="No Instrumentation"
 TITLE[1]="OpenTelemetry No Export"
 TITLE[2]="OpenTelemetry Zipkin"
