### 2026-07-31 2bd16ba
Fix python: Use with for spans
