### 2026-08-01 ab1eec9
Use individual venv for each python framework, use configured $PYTHON from the outside and don't overwrite
