### 2026-08-01 0d50383
Redirect output to logfile
