### 2026-08-01 f3a2af4
Set python and pip correctly for the chosen version
