### 2026-08-01 3eb532b
Fix external python path
