### 2026-08-01 91a96ff
Initial draft for pyspy diagrams
