### 2026-08-01 a816e41
Use configurable PYTHON variable for version
