### 2026-08-01 a36b0e4
Enable pyspy for otel
