### 2026-08-03 d152d81
Automatically create results directory
