### 2026-08-03 9f6906f
Add classifiers after framework name, remove | in table
