### 2026-08-05 c29a325
Fix warning: It might only be a problem if a user has set $RECURSION_DEPTH, but the script varies the depth
