### 2026-08-06 2ead1a5
Update statistics.r
Modified "frameworks/statistic.r" so that read data is no longer overwritten
