### 2026-08-06 53a07f3
Implementation of k-perf Kotlin benchmark
Adds the Kotlin-based MooBench implementation used to measure and compare the runtime overhead introduced by the k-perf compiler plugin.
