### 2026-08-20 981247c
Build SuT-kotlin only with withKotlin flag
SUT-kotlin is built only when the withKotlin flag is set.
