+++ .github/workflows/benchmarkKPerfKotlin.yaml
@@ -36,7 +36,7 @@ jobs:
           chmod 600 ~/.ssh/known_hosts          
       - name: Execute MooBench Benchmark
         run: |
-          ./setup.sh
+          ./setup.sh withKotlin
           cd frameworks/k-perf-kotlin/
           export NUM_OF_LOOPS=10
           ./measure.sh 
