### 2026-08-27 35315e7
Set MOOBENCH_CONFIGURATIONS to get GH actions format
