### 2026-09-04 493ca8f
Additional changes for k-perf-kotlin
Adding the "org.gradle.jvmargs" property
Adding a working "cleanup" function
Starting the build with the flag "--no-deamon"
