### 2026-09-04 0560a44
Additional changes for k-perf-kotlin
Adding the "org.gradle.jvmargs" property
Adding a working "cleanup" function
Starting the build with the flag "--no-deamon"
