+++ frameworks/k-perf-kotlin/functions.sh
@@ -8,10 +8,31 @@ fi
 
 
 function cleanup {
-    [ -f "${BASE_DIR}/hotspot.log" ] && mv "${BASE_DIR}/hotspot.log" "${RESULTS_DIR}/hotspot-${i}-$RECURSION_DEPTH-${k}.log"
-    sync
-    sleep "${SLEEP_TIME}"
-}
+	[ -f "${BASE_DIR}/hotspot.log" ] && mv ...
+
+	mapfile -d '' files_to_delete < <(
+	find "${BASE_DIR}" -type f \
+		\( -name 'trace_*' -o -name 'symbols_*' \) -print0
+	)
+
+	if [ ${#files_to_delete[@]} -eq 0 ]; then
+	echo "No trace_* or symbols_* files found."
+	else
+	echo "Found ${#files_to_delete[@]} trace/symbol file(s):"
+	for file in "${files_to_delete[@]}"; do
+		size_bytes=$(stat -c '%s' "$file")
+		size_human=$(numfmt --to=iec-i --suffix=B "$size_bytes")
+		line_count=$(wc -l < "$file")
+		echo "  Size:  ${size_bytes} bytes (${size_human})"
+		echo "  Lines: ${line_count}"
+	done
+	rm -f -- "${files_to_delete[@]}"
+	echo "Cleanup completed."
+	fi
+
+	sync
+	sleep "${SLEEP_TIME}"
+ }
 
 ## Execute Benchmark
 function executeBenchmark() {
