+++ frameworks/statistics.r
@@ -2,7 +2,7 @@
 # R - script to collect all moobench results
 ############################################
 
-# these values are here only as documentation. The parameters are set by benchmark.sh
+# these values are here only as documentation. The parameters are set by measure.sh
 #rm(list=ls(all=TRUE))
 #data_fn="data/"
 #folder_fn="results-benchmark-binary"
@@ -12,7 +12,7 @@
 #out_yam_fn="results.yaml"
 
 #########
-# These are configuration parameters which are automatically prepended to this file by the benchmark.sh script.
+# These are configuration parameters which are automatically prepended to this file by the measure.sh script.
 # Therefore, they must not be set here. The following lines only serve as documentation.
 #configs.loop=10
 #configs.recursion=c(10)
@@ -29,13 +29,30 @@
 # Process configuration
 
 # divisor 1 = nano, 1000 = micro, 1000000 = milli seconds
-timeUnit <- 1000 
+timeUnit <- 1000
 
-# number of Kieker writer configurations 
+# if TRUE, drop the smallest and largest 0.5% of measurement values of each run
+drop_outliers <- False
+outlierFraction <- 0.005
+
+# number of Kieker writer configurations
 numberOfWriters <- length(configs.labels)
 recursion_depth <- configs.recursion
 
-numberOfValues <- configs.loop*(results.count-results.skip)
+# number of measurement values after warm-up, but before optional outlier removal
+numberOfMeasurementValuesPerRun <- results.count - results.skip
+
+# number of values dropped at each end of each run
+numberOfOutliersPerSide <- if (drop_outliers) {
+   floor(numberOfMeasurementValuesPerRun * outlierFraction)
+} else {
+   0
+}
+
+# number of values actually used for interpretation
+numberOfValuesPerRun <- numberOfMeasurementValuesPerRun - 2 * numberOfOutliersPerSide
+numberOfValues <- configs.loop * numberOfValuesPerRun
+
 numbers <- c(1:(numberOfValues))
 resultDimensionNames <- list(configs.labels, numbers)
 
@@ -47,21 +64,46 @@ resultsBIG <- array(dim=c(numberOfWriters, numberOfValues), dimnames=resultDimen
 
 ## "[ recursion , config , loop ]"
 
-numOfRowsToRead <- results.count-results.skip
+numOfRowsToRead <- numberOfMeasurementValuesPerRun
 
 for (writer_idx in configs.indices) {
    recordsPerSecond = c()
    rpsLastDuration = 0
    rpsCount = 0
-   array_idx <- writer_idx + 1 
-   
+   array_idx <- writer_idx + 1
+
    # loop
    for (loop_counter in (1:configs.loop)) {
       results_fn_filepath <- paste(results_fn, "-", loop_counter, "-", recursion_depth, "-", writer_idx, ".csv", sep="")
       message(results_fn_filepath)
-      results <- read.csv2(results_fn_filepath, nrows=numOfRowsToRead, skip=results.skip, quote="", colClasses=c("NULL","numeric", "numeric", "numeric"), comment.char="", col.names=c("thread_id", "duration_nsec", "gc", "t"), header=FALSE)
-      trx_idx <- (((loop_counter - 1) * numOfRowsToRead + 1):(loop_counter * numOfRowsToRead))
-      resultsBIG[array_idx,trx_idx] <- results[["duration_nsec"]]
+      results <- read.csv2(
+         results_fn_filepath,
+         nrows=numOfRowsToRead,
+         skip=results.skip,
+         quote="",
+         colClasses=c("NULL","numeric", "numeric", "numeric"),
[…]
