### 2026-09-04 197240c
Formatting script, adding parameters and removing "--no-daemon"
