+++ frameworks/k-perf-kotlin/functions.sh
@@ -6,33 +6,32 @@ if [ "${BASH_SOURCE[0]}" -ef "$0" ]; then
     exit 1
 fi
 
-
 function cleanup {
-	[ -f "${BASE_DIR}/hotspot.log" ] && mv ...
+  [ -f "${BASE_DIR}/hotspot.log" ] && mv "${BASE_DIR}/hotspot.log" "${RESULTS_DIR}/hotspot-${i}-$RECURSION_DEPTH-${k}.log"
 
-	mapfile -d '' files_to_delete < <(
-	find "${BASE_DIR}" -type f \
-		\( -name 'trace_*' -o -name 'symbols_*' \) -print0
-	)
+  mapfile -d '' files_to_delete < <(
+    find "${BASE_DIR}" -type f \
+      \( -name 'trace_*' -o -name 'symbols_*' \) -print0
+  )
 
-	if [ ${#files_to_delete[@]} -eq 0 ]; then
-	echo "No trace_* or symbols_* files found."
-	else
-	echo "Found ${#files_to_delete[@]} trace/symbol file(s):"
-	for file in "${files_to_delete[@]}"; do
-		size_bytes=$(stat -c '%s' "$file")
-		size_human=$(numfmt --to=iec-i --suffix=B "$size_bytes")
-		line_count=$(wc -l < "$file")
-		echo "  Size:  ${size_bytes} bytes (${size_human})"
-		echo "  Lines: ${line_count}"
-	done
-	rm -f -- "${files_to_delete[@]}"
-	echo "Cleanup completed."
-	fi
+  if [ ${#files_to_delete[@]} -eq 0 ]; then
+    echo "No trace_* or symbols_* files found."
+  else
+    echo "Found ${#files_to_delete[@]} trace/symbol file(s):"
+    for file in "${files_to_delete[@]}"; do
+      size_bytes=$(stat -c '%s' "$file")
+      size_human=$(numfmt --to=iec-i --suffix=B "$size_bytes")
+      line_count=$(wc -l < "$file")
+      echo "  Size:  ${size_bytes} bytes (${size_human})"
+      echo "  Lines: ${line_count}"
+    done
+    rm -f -- "${files_to_delete[@]}"
+    echo "Cleanup completed."
+  fi
 
-	sync
-	sleep "${SLEEP_TIME}"
- }
+  sync
+  sleep "${SLEEP_TIME}"
+}
 
 ## Execute Benchmark
 function executeBenchmark() {
@@ -56,15 +55,15 @@ function runNoInstrumentationJava {
     # No instrumentation
     k=$1
     info " # ${i}.${RECURSION_DEPTH}.${k} ${TITLE[$k]}"
-	
+
     "${JAVA_BIN}" -cp "${MOOBENCH_JAR_PLAIN}" benchmark.MainKt \
-        --output-filename "${RAWFN}-${i}-${RECURSION_DEPTH}-${k}.csv" \
-        --total-calls "${TOTAL_NUM_OF_CALLS}" \
-        --method-time "${METHOD_TIME}" \
-        --total-threads "${THREADS}" \
-        --recursion-depth "${RECURSION_DEPTH}" \
-        ${MORE_PARAMS} \
-        &> "${RESULTS_DIR}/output_${i}_${RECURSION_DEPTH}_${k}.txt"
+         --output-filename "${RAWFN}-${i}-${RECURSION_DEPTH}-${k}.csv" \
+         --total-calls "${TOTAL_NUM_OF_CALLS}" \
+         --method-time "${METHOD_TIME}" \
+         --total-threads "${THREADS}" \
[…]
