### 2026-09-04 d8d7228
Added comment to statistics.r
